core/domain/merchant/merchant.go
Former-commit-id: 57158008475c6af10c4f72472d3d805197417491

+++ core/domain/merchant/merchant.go
@@ -718,7 +718,7 @@ func (a *accountImpl) TransferToMember1(amount float32) error {
 		if registry.MerchantTakeOutCashFree {
 			conf := a.mchImpl._valRep.GetGlobNumberConf()
 			if conf.TakeOutCsn > 0 {
-				csn := 0
+				csn := float32(0)
 				err = m.GetAccount().ChargeForPresent("返还商户提现手续费", "",
 					csn, member.DefaultRelateUser)
 			}
